console_view: Update builders in fake changes correctly
Where we inject a "fake" change, we then never update the list of builders it
contains. This leads to bugs in the UI rendering since the builders list
might not be the accurate in early calls to the onChange function.

Keep track of fake changes and update those too so the builders list
is maintained.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>