schedulers: docs: pass properties to branch schedulers
Unlike the ForceScheduler, the AnyBranchScheduler does not take a list
of parameters but a simple dict. Add a repos_dict_for_scheduler() helper
that returns a dict containing the {repo,branch}_<repo> keys mapping to
the values in config.py. These are used in the factory definition of the
docs build (see the Git steps).

Also add the missing "yab-helper" to the bitbake-docs-changed scheduler.

Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>